+++ numpy/f2py/tests/test_return_integer.py
@@ -2,13 +2,14 @@ from __future__ import division, absolute_import, print_function
 
 from numpy.testing import *
 from numpy import array
+from numpy.compat import long
 import util
 
 class TestReturnInteger(util.F2PyTest):
     def check_function(self, t):
         assert_( t(123)==123,repr(t(123)))
         assert_( t(123.6)==123)
-        assert_( t(123l)==123)
+        assert_( t(long(123))==123)
         assert_( t('123')==123)
         assert_( t(-123)==-123)
         assert_( t([123])==123)
@@ -34,7 +35,7 @@ class TestReturnInteger(util.F2PyTest):
         assert_raises(Exception, t, {})
 
         if t.__doc__.split()[0] in ['t8','s8']:
-            assert_raises(OverflowError, t, 100000000000000000000000l)
+            assert_raises(OverflowError, t, 100000000000000000000000)
             assert_raises(OverflowError, t, 10000000011111111111111.23)
 
 class TestF77ReturnInteger(TestReturnInteger):
